reorganizando e padronizando os packages do projeto
