--- app/src/main/java/com/renan/digitalspace/apod/view/LoginFragment.kt
@@ -1,37 +1,0 @@
-package com.renan.digitalspace.apod.view
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Button
-import androidx.navigation.Navigation
-import com.renan.digitalspace.R
-
-
-class LoginFragment : Fragment() {
-
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_login, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        val navController = Navigation.findNavController(view)
-
-        view.findViewById<Button>(R.id.button).setOnClickListener {
-            navController.navigate(R.id.epicFragment)
-        }
-
-    }
-
-
-}
