+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".apod.view.MainActivity">
+    tools:context=".MainActivity">
 
     <fragment
         android:id="@+id/navigation_header_container"
