+++ app/src/main/res/navigation/navigation_graph.xml
@@ -2,20 +2,4 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/navigation_graph"
-    app:startDestination="@id/loginFragment">
-    <fragment
-        android:id="@+id/loginFragment"
-        android:name="com.renan.digitalspace.apod.view.LoginFragment"
-        android:label="fragment_login"
-        tools:layout="@layout/fragment_login" >
-        <action
-            android:id="@+id/action_loginFragment_to_epicFragment"
-            app:destination="@id/epicFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/epicFragment"
-        android:name="com.renan.digitalspace.epic.view.EpicFragment"
-        android:label="fragment_epic"
-        tools:layout="@layout/fragment_epic" />
-</navigation>
+    android:id="@+id/navigation_graph" />
